renamed module.css except for navigation
